# openapi/tktsync.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TktSync API",
    "version": "1.0.0",
    "description": "Authoritative TktSync API contract for implemented M3, M4, and M5 Admin and Partner surfaces."
  },
  "paths": {
...
    "/api/v1/admission/scans": {
      "post": {
        "operationId": "admissionValidateAndAdmit",
        "security": [
          {
            "ScannerBearer": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/XRequestID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdmissionScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authoritative admission decision. Rejections are persisted ScanAttempt results.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionScanResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "PartnerBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "ScannerBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "SelectionBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque, time-bounded BuyerSelectionSession capability."
      }
    },
    "parameters": {
      "XRequestID": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "Optional client request/correlation UUID.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Durable idempotency key scoped to the authenticated actor and operation.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "Machine-readable error response.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/XRequestID"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message", "request_id", "details"],
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "request_id": {
                "type": "string",
                "format": "uuid"
              },
              "details": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        }
      },
...
      "AdmissionScanRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["event_id", "credential"],
        "properties": {
          "event_id": {
            "type": "string",
            "pattern": "^evt_"
          },
          "credential": {
            "type": "string",
            "minLength": 1,
            "writeOnly": true
          },
          "gate_reference": {
            "type": "string"
          }
        }
      },
      "TicketDisplay": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "section": {
            "type": "string"
          },
          "row": {
            "type": "string"
          },
          "seat": {
            "type": "string"
          }
        }
      },
      "AdmissionScanResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["result", "scan_attempt_id"],
        "properties": {
          "result": {
            "type": "string",
            "enum": [
              "ADMITTED",
              "TICKET_ALREADY_ADMITTED",
              "TICKET_INVALID",
              "TICKET_VOID",
              "CREDENTIAL_REVOKED",
              "CREDENTIAL_SUPERSEDED",
              "WRONG_EVENT",
              "EVENT_CANCELLED",
              "ADMISSION_NOT_OPEN",
              "MANUAL_OVERRIDE_ADMITTED"
            ]
          },
          "scan_attempt_id": {
            "type": "string",
            "pattern": "^scan_"
          },
          "admission_id": {
            "type": "string",
            "pattern": "^adm_"
          },
          "ticket": {
            "type": "object",
            "additionalProperties": false,
            "required": ["id", "display"],
            "properties": {
              "id": {
                "type": "string",
                "pattern": "^tkt_"
              },
              "display": {
                "$ref": "#/components/schemas/TicketDisplay"
              }
            }
          },
          "admitted_at": {
            "type": "string",
            "format": "date-time"
          },
          "previous_admission": {
            "type": "object",
            "additionalProperties": false,
            "required": ["admitted_at", "gate_reference"],
            "properties": {
              "admitted_at": {
                "type": "string",
                "format": "date-time"
              },
              "gate_reference": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "headers": {
      "XRequestID": {
        "description": "Request/correlation identifier generated or normalized by TktSync.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "CacheControlNoStore": {
        "description": "Reservation responses containing protected state must not be cached.",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      }
    }
  }
}